wlan: Indicate disconnect to SME to abort pre-auth
If connect is received when SME is already connected, HDD try to
disconnect the current connection and post disconnect to SME.
If preauth is in progress the disconnect may get timeout and
once again new connect is received which removes disconnect cmd
from SME queue and tries to connect to new AP. This causes
firmware to crash as BSS is already present for the session.

To fix this indicate disconnect to SME so that preauth is
aborted as soon as possible.

Change-Id: I6c3ed5344950cb5acfe9df82acaeb77f605e1d28
CRs-Fixed: 2019891